+++ packages/brick_sqlite/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+## 3.1.1
+
+- Expose a generic type for `MemoryCacheProvider` models
+- Expose a generic type for `SqliteProvider` models
+
 ## 3.1.0
 
 - Apply standardized lints