+++ packages/brick_supabase/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## 1.0.5
+
+- Add `onConflict` to `SupabaseProvider#upsert` requests (#467) @sonbs21
+
+## 1.0.4
+
+- Expose private methods `SupabaseProvider#upsertByType` and `SupabaseProvider#recursiveAssociationUpsert` as protected methods
+
 ## 1.0.3
 
 - Remove `select` from `#delete`